Remove time tree
Date inference is likely inaccurate for this pathogen, and so this commit ensures that a time tree is not visualized in auspice. This is accomplished by removing all node attributes except divergence branch lengths (which are listed in  the `mutation_length` attribute) from branch_lengths.json.
Follows https://github.com/nextstrain/measles/pull/34