Remove FH cluster functionality
Initial development for the tb workflow was conducted on the Fred Hutch cluster, which required:

1. A slurm config
2. fasterq-dump could only be used if fastq files were first downloaded to local scratch space (stored in TMPDIR) as described here: https://sciwiki.fredhutch.org/scicompannounce/2024-02-26-fasterq-dump-issue/

After modifying the workflow to run with a Docker image (previously it had used conda environments), the workflow produced errors that the TMPDIR doesn't have enough space, and so further code changes would be needed for running this workflow on the cluster with Docker.

For now we have decided we don't need the Fred Hutch cluster functionality, and so I'm removing these features.