Delete phylogenetic directories
We decided to use the phylogenetic_vcf workflow rather than the phylogenetic_fasta workflow, and so the phylogenetic_fasta directory can be removed.

Also, we have combined the ingest and phylogenetic workflows (currently located in the ingest directory), and so the phylogenetic_vcf directory can also be removed. The next commit will move the ingest+phylo workflow that is in the ingest directory to the top level of the github repo, and then delete the empty ingest directory.