Ignore SnpEff 5.3.0a
One of the Bioconda builds of this version did not properly declare a
dependency of Java version >= 21. This has been addressed with a newer
build of the same version, but the Conda solver still prefers the broken
build.

Since there isn't a way¹ to specifically exclude the bad build of this
version, the next best solution is to exclude the version entirely.

¹ https://stackoverflow.com/q/77097829